Code Review Changes
Updates from pull request code review
