+++ src/devices/LiquidLevelSwitch/LiquidLevelSwitch.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
@@ -10,6 +10,6 @@
     <Compile Remove="samples\**" />
     <None Include="README.md" />
 	  <ProjectReference Include="$(MainLibraryPath)System.Device.Gpio.csproj" />
-	  <!--<ProjectReference Include="..\Common\Iot\Device\Common\CommonHelpers.csproj" />-->
   </ItemGroup>
+  
 </Project>
