+++ src/devices/LiquidLevelSwitch/PinMode.cs
@@ -1,0 +1,22 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Iot.Device.LiquidLevelSwitch
+{
+    /// <summary>
+    /// Liquid level liquid detection pin mode
+    /// </summary>
+    public enum PinMode
+    {
+        /// <summary>
+        /// Liquid present pin mode is low
+        /// </summary>
+        Low = 0,
+
+        /// <summary>
+        /// Liquid present pin mode is high
+        /// </summary>
+        High = 1
+    }
+}
