+++ src/devices/Llc200d3sh - Copy/README.md
@@ -1,0 +1,11 @@
+# Optomax digital liquid level switch
+
+Optomax digital liquid level switch is a device that can detect the presence of liquid/water. GPIO can be used to communicate with the device.
+
+[Datasheet](https://cdn-shop.adafruit.com/product-files/3397/3397_datasheet_actual.pdf) for the Optomax digital liquid level switch.
+
+An example on how to use this device binding is available in the [samples](samples) folder.
+
+The following fritzing diagram illustrates one way to wire up the Optomax digital liquid level switch with a Raspberry Pi.
+
+![Raspberry Pi Breadboard diagram](samples/rpi-llc200d3sh_bb.png)
