+++ src/devices/Llc200d3sh - Copy/samples/Llc200d3sh.Samples.csproj
@@ -1,0 +1,23 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <EnableDefaultItems>false</EnableDefaultItems>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="rpi-llc200d3sh.fzz" />
+    <None Remove="rpi-llc200d3sh_bb.png" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <Compile Include="Llc200d3sh.sample.cs" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <ProjectReference Include="..\Llc200d3sh.csproj" />
+    <ProjectReference Include="$(MainLibraryPath)System.Device.Gpio.csproj" />
+  </ItemGroup>
+  
+</Project>
