--- src/devices/Llc200d3sh/DetectionType.cs
@@ -1,18 +1,0 @@
-﻿namespace Iot.Device.Llc200d3sh
-{
-    /// <summary>
-    /// Optomax digital liquid level switch detection types
-    /// </summary>
-    public enum DetectionType
-    {
-        /// <summary>
-        /// Liquid detected
-        /// </summary>
-        Liquid = 0,
-
-        /// <summary>
-        /// Air detected
-        /// </summary>
-        Air = 1
-    }
-}
